problem: Hide icon in secondary menu should have disabled color
